Remove blog_example.rs and data.rs files; refactor request module to use request_data.rs instead. This cleanup eliminates unused code and improves module organization.
